# backend-flask/api/open_api/openapi.yaml
openapi: 3.0.3
info:
  title: "SC API pos"
  description: "Модуль для интеграции с кассами для лояльности на смарт контрактах."
  version: 1.0.0
servers:
  - url: https://dm.api/20130701/
    description: "Production сервер"
  - url: http://staging-api.example.com
    description: "Staging сервер"
...
paths:
...
  /loyalties/{merchant_id}:
    get:
      summary: "Поиск лояльности"
      operationId: src.loyalties.views.get_loyalty
      tags:
      - Loyalties
      description: "Получение полной информации о ПЛ (Программы Лояльности) участника Системы Rusbonus. Программа лояльности может отсутствовать, а также содержать нулевые значения порогов и скидок\n\n."
      parameters:
      - $ref: '#/components/parameters/DM-Authorization'
      - $ref: '#/components/parameters/_dmapptoken'
      - $ref: '#/components/parameters/merchant_id'
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Loyalty'
...
  /adminsc/loyalty/subscription:
    get:
      summary: "Получить статус подписки"
      operationId: src.adminsc.views.get_subscription
      tags:
      - AdminSC
      description: "Получить статус подписки"
      security:
        - bearerAuthAdmin: []
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                properties:
                  isActive:
                    type: boolean
                    description: "Подписка активна или нет"
                    example: true
  /adminsc/loyalty:
    get:
      summary: "Информация о лояльности"
      operationId: src.adminsc.views.get_loyalty
      tags:
      - AdminSC
      description: "Информация о лояльности"
      security:
        - bearerAuthAdmin: []
      responses:
        '200':
          description: ""
          content:
            application/json:
              schema:
                properties:
                  accrualPercent:
                    type: string
                    description: "курс начисления бонусных баллов"
                    example: "1"
                  minAccrualThreshold:
                    type: string
                    description: "минимальная сумма покупки, при которой производится начисление бонусных баллов"
                    example: "1"
                  accrualRate:
                    type: string
                    description: "курс списания бонусного балла"
                    example: "1"
                  maxWriteoffPercent:
                    type: string
                    description: "процент покупки, который может быть оплачен за счет бонусов"
                    example: "100"
                  bonusActivationDelay:
                    type: string
                    description: "срок активации бонусных баллов, сек."
                    example: "400"
                  bonusLifetime:
                    type: string
                    description: "срок жизни бонусных баллов, сек."
                    example: "500"
                  name:
                    type: string
                    description: "Тип контракта."
                    example: "Bonus"
                  generation:
                    type: string
                    description: "Версия контракта"
                    example: "3.0"
                  isSimultaneous:
                    type: boolean
                    description: "одновременное списание и начисление баллов."
                    example: true
              examples:
                v3:
                  value:
                    accrualPercent: "2"
                    minAccrualThreshold: "1"
                    accrualRate: "1"
                    maxWriteoffPercent: "100"
                    isSimultaneous: true
                    name: "Bonus"
                    generation: "3.0"
                v2:
                  value:
                    accrualPercent: "2"
                    minAccrualThreshold: "1"
                    accrualRate: "1"
                    maxWriteoffPercent: "100"
                    bonusActivationDelay: "400"
                    bonusLifetime: "500"
                    isSimultaneous: true
                    name: "Bonus"
                    generation: "2.0"

    put:
      summary: "Редактировать лояльность"
      operationId: src.adminsc.views.edit_loyalty
      tags:
      - AdminSC
      description: "Редактировать лояльность"
      security:
        - bearerAuthAdmin: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                accrualPercent:
                    type: string
                    description: "курс начисления бонусных баллов"
                    example: "100"
                minAccrualThreshold:
                  type: string
                  description: "минимальная сумма покупки, при которой производится начисление бонусных баллов"
                  example: "1"
                accrualRate:
                  type: string
                  description: "курс списания бонусного балла"
                  example: "1"
                maxWriteoffPercent:
                  type: string
                  description: "процент покупки, который может быть оплачен за счет бонусов"
                  example: "100"
                bonusActivationDelay:
                  type: string
                  description: "срок активации бонусных баллов, сек."
                  example: "1"
                bonusLifetime:
                  type: string
                  description: "срок жизни бонусных баллов, сек."
                  example: "999999"
                isSimultaneous:
                  type: boolean
                  description: "одновременное списание и начисление баллов."
                  example: true

      responses:
        '202':
          description: ""
          content:
            {}
        '402':
          description: "Требуется оплата. Подписка не актива"
          content:
            application/json:
              schema:
                example: {}
components:
  securitySchemes:
    bearerAuthEvotor:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_evotor

    bearerAuthAdmin:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: src.middleware.bearer_auth_adminsc

    bearerAuthPos:
      type: apiKey
      in: header
      name: Dmpos-Authorization
      x-apikeyInfoFunc: src.middleware.bearer_auth_pos
...
  parameters:
    DM-Authorization:
      name: "DM-Authorization"
      in: header
      schema:
        type: string
        example: "dmapptoken 550e8400-e29b-41d4-a716-446655440000"
        description: "Этот ключ создаётся и выдаётся разработчикам приложений, использующим API, сотрудниками компании Rusbonus."
    _dmapptoken:
      name: "_dmapptoken"
      in: query
      schema:
        type: string
        example: "dmapptoken 550e8400-e29b-41d4-a716-446655440000"
        description: "Этот ключ создаётся и выдаётся разработчикам приложений, использующим API, сотрудниками компании Rusbonus."
    merchant_id:
      name: "merchant_id"
      in: path
      required: true
      schema:
        type: string
        maxLength: 100
...
  schemas:
...
    Loyalty:
      type: object
      properties:
        type:
          type: string
          example: "bonus"
        currency_code:
          type: integer
          format: int64
          example: 643
          description: "Целое положительное число\n\nЦифровой код валюты в стандарте ISO 4217 [http://ru.wikipedia.org/wiki/ISO_4217]."
        currency_name:
          type: string
          example: "RUB"
          description: "Строка символов длиной 3.\n\nАлфавитный код валюты в стандарте ISO 4217 [http://ru.wikipedia.org/wiki/ISO_4217]."
        min_purchase_amount:
          type: integer
          format: int64
          example: 1
          description: "Положительное целое число.\n\nМинимальная сумма покупки для начисления бонусных баллов."
        max_purchase_percentage:
          type: integer
          format: int64
          example: 50
          description: "Положительное целое число в диапазоне от 0 до 100.\n\nМаксимальная часть стоимости покупки в процентах, которая может быть оплачена бонусными баллами."
        expiration:
          type: integer
          format: int64
          example: 30
          description: "Положительное целое число.\n\nКоличество дней, спустя которые начисленные бонусные баллы сгорят."
        amount_to_bonus:
          type: array
          description: "Список алфавитно-цифровых строк без знаков препинания, разделённых запятыми.\n\nСписок идентификаторов купонов (сертификатов) пользователя, которые следует погасить в этой покупке."
          items:
            type: string
          example: [100, "10.0000"]
        bonus_to_amount:
          type: array
          description: "Список алфавитно-цифровых строк без знаков препинания, разделённых запятыми.\n\nСписок идентификаторов купонов (сертификатов) пользователя, которые следует погасить в этой покупке."
          items:
            type: string
          example: [1, "0.1000"]
...